# doc/api.yaml
openapi: 3.1.0
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!

    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
components:
  schemas:
    User:
      title: User
      description: This object represents a single user
      type: object
      properties:
        userId:
          type: integer
          description: Unique identifier for the user
          example: 1234
          readOnly: true
          minimum: 0
        username:
          description: The username displayed to the others
          type: string
          example: CoolUsername42
          minLength: 3
          maxLength: 32
          pattern: '^[A-Za-z0-9_\-]{3,32}$'
      required: [ "userId", "username" ]
    Profile:
      title: Profile
      description: This object represents the profile data of a user
      type: object
      properties:
        userId: { $ref: '#/components/schemas/User/properties/userId' }
        username: { $ref: '#/components/schemas/User/properties/username' }
        photoCount:
          description: The count of how many photos this user uploaded
          type: integer
          example: 42
          minimum: 0
        followersCount:
          description: The amount of users following this user
          type: integer
          example: 42
          minimum: 0
        followingCount:
          description: The amount of users followed by this user
          type: integer
          example: 42
          minimum: 0
        bannedCount:
          description: The amount of users banned by this user
          type: integer
          example: 42
          minimum: 0
      required: [ "userId", "username", "photoCount", "followersCount", "followingCount", "bannedCount" ]
...
    Photo:
      title: Photo
      description: This object represents a single photo in the user's feed
      type: object
      properties:
        photoId:
          type: integer
          description: Unique identifier for the photo
          example: 1234
          minimum: 0
          readOnly: true
        ownerId: { $ref: '#/components/schemas/User/properties/userId' }
        ownerUsername: { $ref: '#/components/schemas/User/properties/username' }
        image:
          type: string
          format: binary
          description: Binary data of the image
          minLength: 0
          maxLength: 999999999 # This is here to handle warnings, it's not a real limit
        mimeType:
          type: string
          description: The MIME type of the image
          example: "image/jpeg"
          pattern: '^[A-Za-z0-9_\-]{3,}$'
          minLength: 3
          maxLength: 32
        caption:
          type: string
          description: Caption for the image
          example: "A beautiful landscape"
          pattern: '^[\p{L}\p{N}\p{M}\p{P}\p{S} ]{0,32}$'
          minLength: 0
          maxLength: 32
        uploadTime:
          type: string
          format: date-time
          description: The date and time at which this photo was uploaded, RFC 3339 format
          example: 2017-07-21T17:32:28Z
        likeCount:
          type: integer
          description: Number of likes for the image
          example: 42
          minimum: 0
        commentsCount:
          type: integer
          description: Number of comments for the image
          example: 42
          minimum: 0
      required: [ "photoId", "ownerId", "image", "uploadTime", "likeCount", "commentsCount" ]
    Error:
      title: Error
      description: |-
        This object represents a single error to be returned for certain requests
      type: object
      properties:
        code:
          type: integer
          description: An HTTP error code
          example: 404
        message:
          type: string
          description: A brief message describing the error code
          example: The specified resource was not found
      required: [ "code", "message" ]
  parameters:
    usernameParam:
      name: username
      in: path
      description: Username of the user
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/username' }
    targetUsernameParam:
      name: targetUsername
      in: path
      description: Username of the target
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/username' }
    photoIdParam:
      name: photoId
      in: path
      description: ID of the photo
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Photo/properties/photoId' }
    likeIdParam:
      name: likeId
      in: path
      description: ID of the like (that being the id of the liker)
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/userId' }
...
  responses:
...
    NoContent:
      description: The request was successful, but there is no content to return.
    # 400
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 401
    Unauthorized:
      description: Invalid or expired authentication credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 404
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 409
...
    UnprocessableEntity:
      description: The request was well-formed but was unable to be followed due to semantic errors.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [ ]
...
paths:
  /session:
    post:
      tags: [ "Login" ]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created, and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: login schema
              properties:
                username: { $ref: '#/components/schemas/User/properties/username' }
      responses:
        "201":
          description: User log-in action successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
      security: [ ]
...
  /users/{username}/profile:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "User" ]
      operationId: getUserProfile
      summary: Retrieve the information of a user
      description: |-
        Given a user's username, retrieve all the public info available.
      responses:
        "200":
          description: Successfully retrieved profile data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/photos:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Photo" ]
      operationId: getMyPhotos
      summary: Retrieve the photos of the user
      description: |-
        Given a user's username, retrieve all the photos uploaded by them.
      responses:
        "200":
          description: Successfully retrieved photos
          content:
            application/json:
              schema:
                type: array
                description: list of photos
                items: { $ref: '#/components/schemas/Photo' }
                uniqueItems: true
                minItems: 0
                maxItems: 99999
          "400": { $ref: '#/components/responses/BadRequest' }
          "401": { $ref: '#/components/responses/Unauthorized' }
          "404": { $ref: '#/components/responses/NotFound' }
          "500": { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [ "Photo" ]
      operationId: uploadPhoto
      summary: Create a post with the given photo
      description: |-
        Given the file of an image, upload it to the server under a new Photo.
      requestBody:
        description: The image to upload along with an optional description
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: upload schema
              properties:
                image: { $ref: '#/components/schemas/Photo/properties/image' }
                mimeType: { $ref: '#/components/schemas/Photo/properties/mimeType' }
                caption: { $ref: '#/components/schemas/Photo/properties/caption' }
      responses:
        "201":
          description: Photo uploaded successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/photos/{photoId}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/photoIdParam'
    delete:
      tags: [ "Photo" ]
      operationId: deletePhoto
      summary: Delete a photo from the current user
      description: |-
        Given the id of a photo from the current user's posts, delete it.
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/photos/{photoId}/likes:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/photoIdParam'
    post:
      tags: [ "Like" ]
      operationId: likePhoto
      summary: Like a photo
      description: Given a user's username, and a photo's id, like it.
      responses:
        "201":
          description: Photo liked successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/photos/{photoId}/likes/list:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/photoIdParam'
    get:
      tags: [ "Like" ]
      operationId: getPhotoLikers
      summary: Retrieve the IDs of the users that liked a photo
      description: |-
        Given the id of a photo from a user, retrieve all the IDs of the users that liked it.
      responses:
        "200":
          description: Successfully retrieved likers
          content:
            application/json:
              schema:
                type: array
                description: list of likes
                items: { $ref: '#/components/schemas/User/properties/userId' }
                uniqueItems: true
                minItems: 0
                maxItems: 99999
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/photos/{photoId}/likes/list/{targetUsername}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/photoIdParam'
      - $ref: '#/components/parameters/targetUsernameParam'
    get:
      tags: [ "Like" ]
      operationId: getLikeStatus
      summary: Check if a user liked a photo
      description: |-
        Given a user's username, check if they liked a photo.
      responses:
        "200":
          description: User liked the photo
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasLiked:
                    type: boolean
                    description: Whether the user liked the photo
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/photos/{photoId}/likes/{likeId}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/photoIdParam'
      - $ref: '#/components/parameters/likeIdParam'
    delete:
      tags: [ "Like" ]
      operationId: unlikePhoto
      summary: Unlike a photo
      description: Given a user's username and a photo's id, unlike it.
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...